arm: Introduce and use arm_dcaches_flush
This generalizes arch_paging_flush_cpu_caches to arm_dcaches_flush, a
dcache flusher by MVA that can be instructed to either clean, invalidate
or do both on a memory region mapped to the hypervisor address space.

Already arch_paging_flush_cpu_caches was too large to be inlined, so
arm_dcaches_flush is implemented in mmu_hyp.c. This version also takes
regions into account that are not multiples of the minimum cache line
size.

CC: Marc Zyngier <marc.zyngier@arm.com>
CC: Mark Rutland <mark.rutland@arm.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>