arm: Implement arm_cell_dcaches_flush
This flushes all dcache entries related to a specific cell by mapping
each physical RAM page of the cell into the hypervisor and then perform
the requested flush on the corresponding virtual address. Those flushes
will be broadcast to all CPUs, thus the call only needs to be performed
once on any CPU in the system.

This pattern was bluntly stolen from KVM. It will serve as a building
block to emulate guest-issued set/way cache maintenance operations.

CC: Marc Zyngier <marc.zyngier@arm.com>
CC: Mark Rutland <mark.rutland@arm.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>