arm: Refactor SMP setup and cleanup
Prepare for a cleaner setup and cleanup of the SMP subsystem: introduce
a platform init function called once during setup as well as per-cell
init/exit functions. Those are stubbed for PSCI-based platforms and only
need to be implemented for vexpress so far.

We emulate the previous invocation of the init op for the root cell
during setup via smp_init. Note that init was and is only called for the
root cell (a highly confusing interface).

One difference of the new code for vexpress is that we will now expose
the mailbox interface for kicking off secondary CPUs also to non-root
cells. So far we only offered PSCI emulation, but there is no point in
hiding the FLAGSSET register in an emulated for from them - they will
now be able to chose the wake-up method.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>