arm: Fix root cell irqchip bitmap manipulations
We always operated only on the first 32 pins, relative to an the config
structure that was applied.

Shouldn't have caused problems so far because most non-root configs
assign only small interrupt sets.

Fixes f2931b157d70.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>

In addition, this is the change log that was forgotten to add to
commit f2931b157d70:

arm: Use full bitmap to manage per-cell interrupt access

Extend the cell irqchip bit bitmap to 1024 pins, exploiting the
clustered irqchip configuration format. We require each irqchip config
entry to be 32-pins aligned to each to merging of into the cell's
runtime bitmap.

Pins 0..31 minus hypervisor-used SGIs and PPIs remain freely accessible
to the cells because those are per-CPU.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>