x86: Rework ioapic_get_or_add_phys
ioapic_get_or_add_phys currently returns NULL if anything went wrong,
which not only includes lack of memory but also an overflow of the
number of registered physical IOAPICs. Improve the interface of this
function by returning an error code directly and the actual pointer
via an argument pointer.

Already make it public, it will also be used by vtd.c.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>