x86: Extend number of support IOAPIC pins to 120
There are actually IOAPICs out there with more than the usual 24 pins.
Specs like the one of the ICH9 state that there can be up to 240 pins
per IOAPIC. That would mean we would have to use up to two irqchip
config entries per IOAPIC. However, those specs do not tell how pins
beyond 120 should be addressed. So let's save this refactoring exercise
for the time when this is clear and simply lift the limit to 120 -
which is also still manageable with a single entry.

The code now evaluates the number of pins in ioapic_get_or_add_phys,
rejects too large numbers, and otherwise takes the dynamic limit into
account when managing the redirection entries or reserving IR entries
for the IOAPIC.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>