core: ivshmem: Mark device as different via subsystem vendor ID
The is an increasing difference between the original IVSHMEM device (as
implemented by QEMU) and our version, still counting. Until we have
come to a conclusion how to move on with the device ("IVSHMEM 2.0" vs.
something new), mark our device as different via the PCI subsystem ID.
This allows our drivers to find the right devices.

We are using the Siemens vendor ID, but this is nothing official, just
a natural local choice.

CC: Henning Schild <henning.schild@siemens.com>
CC: Måns Rullgård <mans@denx.de>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>